Add monospace formatting to code fragments
The general strategy is to take the raw HTML such as

blah <code>x = 1</code> cats

and map it to

blah <code>[[&lt;[x = 1]&gt;]]</code> cats

Then apply Html.fromHtml, which produces

blah [[<[x = 1]>]] cats

Using these tokens, the text can be parsed and a monospace font can
be applied.

The transform of the CharSequence to a string in the setCodeFont
method is to simplify the process of search for the tokens, though
it could probably be implemented using the mutable nature of
SpannableStringBuilder. Performance wise there is probably not a huge
hit since we are parsing relatively small amounts of text.