Fix crash when all subreddit posts are loaded
When the app loaded all posts for a subreddit, SubredditPosts.nomore
is set to true, which in turn alters the output of SubmissionAdapter.getItemCount
to just be {number of posts} (not {number of posts + 1} or 0).

However, the ViewHolder with the loading spinner still exists in the
RecyclerView, so the exception is thrown since our reported number of
items is {number of posts}, but the RecyclerView has {number of posts + 1}
items.

This solution adds the nomoreposts.xml which replaces loading spinner
with a "No more posts!" message, and alters the SubmissionAdapter.getItemCount
method return either 0 or posts.size() + 1 (so it always accounts for
footer element).

Some refactoring was also applied to make code a bit more readable.