Add spoiler functionality
This has been tested on subreddits that use
- #s
- /s
- /sp
- /spoiler

as spoiler identifiers. The effect is that the visible text is
displayed as a link, with a whitespace, then the spoiler text
blacked out. When the visible text the clicked, the text of the
spoiler text is reset to white, with the backgroud to stay black.

This is a bit fragile due to the fact that:
1. subreddits use different hrefs for spoiler tags,
2. the visible spoiler text is different per subreddit (and even
   per post).

Accordingly, there are a few assumptions. Firstly:

A1. Spoilers are identified by the presence of the "title" attribute
    within the <a> tag.

This seems consitent with the html that comes out of reddit, and
is used in the regex to find each spoiler.

A2. The text within the spoiler does not end in "]]".

The whole parsing token is "]s]]", but this is relaxed to just "]]"
since it may or may not happen. Should probably be strengthed in
the future.

A3. If the title attribue is empty, insert "spoiler" as the visible
    text.

This is fairly self-explanatory, since having no title will mean
there is nothing to click on.

If more patterns need to be added to identify spoilers, see
ImageType.getImageType(String url).

To prevent collapsing of the children tree when a spoiler link is pressed,
ActiveTextView.spoilerClicked is set to true and is checked in the onClick
function in CommentAdapter for the R.id.content element.

Some amount of code clean up was added, since there were a lot of
inconsitent spacing issues and a few indents which could be reduced
by re-wording if statements.