Fix deadend when swiping through posts
The principle issue is CommentScreen does not have access to the
SubredditPosts object, so it does not possess the ability to load
more submissions.

This part of the fix moves Fragment specific logic from SubredditPosts/
MultiredditPosts to the SubmissionDisplay interface, which provides
callback methods which are called when the submissions have been updated.

Since CommentScreen is used for multireddits as well, the PostLoader
interface provides the neccesary methods to load submissions.

There are also numerous refactorings/removal of dead code, most of
which are minor.