Improve seeking in MediaVideoView
Use the new seekTo(long, int) on API>=26. This gives finer control over
the frame to seek to. On non low-ram devices, SEEK_CLOSEST is used to
seek to the nearest frame. On low-ram devices, SEEK_CLOSEST_SYNC is used
to seek to the nearest key-frame. This is an improvement over
seekTo(long) (which uses SEEK_PREVIOUS_SYNC) in both cases.