Fix crosspost crash on startup
The logic flow before was:
  Check if json does not have crosspost_parent_list AND
  get member 0 of json's crosspost_parent_list to check if it's not null

The problem is if the first condition is true(json does not have crosspost_parent_list),
AND can't short-circut and it will move on to evaluate the second side, so it will
try and get member 0 of a crosspost_parent_list, which will be null as we figured out
in the first conditional.

Now the logic flow is:
  Check if json does not have crosspost_parent_list OR
  get member 0 of json's crosspost_parent_list to check if it's null

This way, the OR can short circut on the first true condition, so the get member 0
logic won't trigger unless the json does have crosspost_parent_list.