Partial fix for keyboard not focusing on Pie
I hate Pie more and more with each day. Pie does something fucky with
focus and the keyboard - as best I can tell, the user taps the reply
button, we focus the textbox, then the keyboard steals focus before
anything can happen so the textbox doesn't move up or gain real focus.
The keyboard *does* give focus back to the app once it's popped up, but
Pie isn't handling that focus correctly, it's more like fake focus.

This is a hacky, gross, really absolutely horrible workaround, and
doesn't always work, but it fixes the most annoying instances of the
issue.